connected to backend, now gonna make loading screen and backend put into postgres
